+++ streams/src/test/java/org/apache/kafka/streams/processor/internals/StreamThreadTest.java
@@ -715,6 +715,7 @@ public class StreamThreadTest {
         EasyMock.replay(consumer, consumerGroupMetadata);
 
         final AtomicBoolean committed = new AtomicBoolean(false);
+        final TopologyMetadata topologyMetadata = new TopologyMetadata(internalTopologyBuilder, config);
         final TaskManager taskManager = new TaskManager(
             null,
             null,
@@ -723,8 +724,7 @@ public class StreamThreadTest {
             null,
             null,
             null,
-            null,
-            null,
+            topologyMetadata,
             null,
             null
         ) {
@@ -736,7 +736,6 @@ public class StreamThreadTest {
                 return 1;
             }
         };
-        final TopologyMetadata topologyMetadata = new TopologyMetadata(internalTopologyBuilder, config);
         topologyMetadata.buildAndRewriteTopology();
         final StreamThread thread = buildStreamThread(consumer, taskManager, config, topologyMetadata);
 
@@ -797,7 +796,6 @@ public class StreamThreadTest {
             activeTaskCreator,
             standbyTaskCreator,
             topologyMetadata,
-            null,
             null,
             null
         ) {
